looprpc: fractional amount for static address loop-ins
The QuoteRequest adds a field select_deposits to signal
that the specified quote amount should be coin-selected from the client's deposits in order to determine the
number of deposits to quote for.
The StaticAddressLoopInRequest gets a new amount field that either indicates that the swap amount should be allocated from the passed deposits or coin-selected from all available deposits.